requirements.txt
pytesseract>=0.3.10
PyMuPDF>=1.23.0
Pillow>=9.4.0
PyPDF2>=3.0.0
tqdm>=4.65.0